Cleanup the mic example, remove the test.
